+++ src/events/spawn_animated_sprite_event.rs
@@ -4,8 +4,8 @@ use super::spawn_sprite_event::SpawnSpriteEvent;
 
 #[derive(Event)]
 pub struct SpawnAnimatedSpriteEvent {
-    pub tile_size: f32,
-    pub tile_columns: usize,
+    pub tile_size: u32,
+    pub tile_columns: u32,
     pub frame_timing: f32,
     pub frame_count: usize,
     pub spawn_sprite_event: SpawnSpriteEvent,
